* Major Change: 	removed recursion from stringToBytesHelper. This was causing a segmentation fault due to too many recursive calls. 	replaced with a standard 'for-loop'

+++ characterFrequency/characterFrequency.cpp
@@ -7,9 +7,9 @@
 using namespace std;
 
 
-map<char, int> get_char_freq (ifstream& file) {
+map<char, long> get_char_freq (ifstream& file) {
     char chr;
-    map<char, int> charFreq;
+    map<char, long> charFreq;
     while (file.get(chr))
         charFreq[chr]++;
     charFreq[EOF] = 1; // char to rep End of File
